Update the OUTLINE per Surya's comments
Surya's comments:

I think most interest will be in 2-pt statistics before we get to
MKS. So I suggest we organize in increasing order of interest (and
complexity) as (i) 2 pt. statistics and PCA, (ii) example on elastic
homogenization fit using polynomials, and (iii) example on elasticity
MKS (using the same database as (ii)). This should be plenty for the
workshop....